abstract out file conversion
