+++ src/lib/util/mediaLoader.ts
@@ -1,37 +1,8 @@
 import { get } from "svelte/store";
 import { studio } from "../stores";
 import { v4 as uuidv4 } from "uuid";
-import { parseMIME } from "./mimeParser";
-import { assertBrowserSupportsContainer } from "./browserParser";
-import { ffmpegInstance } from "./FFmpegManager";
-import { fetchFile } from "@ffmpeg/ffmpeg";
 
-export const loadMediaMetadata = async (file: File): Promise<App.VideoMedia | App.AudioMedia | App.ImageMedia> => {
-  const MIME = await parseMIME(file);
-  if(MIME === "file/unknown") throw new Error(`Unsupported container: ${MIME}`);
-
-  let src: string;
-
-  if (!await assertBrowserSupportsContainer(MIME)) {
-    const mimeType = MIME.split("/")[0], conversionExt = mimeType === "video" ? "mp4" : mimeType === "audio" ? "mp3" : "jpg";
-    console.log(`Browser does not support ${MIME} container; converting to ${conversionExt}`);
-
-    ffmpegInstance.FS("writeFile", file.name, await fetchFile(file));
-
-    ffmpegInstance.FS("writeFile", `ffmpeg.${conversionExt}`, "");
-    await ffmpegInstance.run("-i", file.name, "-codec", "copy", `ffmpeg.${conversionExt}`);
-
-    const data = ffmpegInstance.FS("readFile", `ffmpeg.${conversionExt}`);
-    const blob = new Blob([data.buffer], { type: `${mimeType}/${conversionExt}` });
-
-    src = URL.createObjectURL(blob);
-
-    ffmpegInstance.FS("unlink", file.name);
-    ffmpegInstance.FS("unlink", `ffmpeg.${conversionExt}`);
-  } else  {
-    src = URL.createObjectURL(file);
-  }
-
+export const parseMediaMetadata = async (file: File, src: string): Promise<App.VideoMedia | App.AudioMedia | App.ImageMedia> => {
   const defaultMediaProperties = {
     uuid: uuidv4(),
     src,
@@ -70,7 +41,7 @@ export const loadMediaMetadata = async (file: File): Promise<App.VideoMedia | Ap
   } else throw Error("Unsupported file type");
 };
 
-export const loadMediaDuration = (src: string, type: string) => new Promise<number>((resolve, reject) => {
+const loadMediaDuration = (src: string, type: string) => new Promise<number>((resolve, reject) => {
   if(type === "audio") {
     const audio = document.createElement("audio");
     audio.src = src;
@@ -98,7 +69,7 @@ export const loadMediaDuration = (src: string, type: string) => new Promise<numb
  * Parses a media source and returns a series of thumbnails to use as a preview.
  * TODO: implement zoomScale as factor in determining how many thumbnails to generate.
  */
-export const loadThumbnails = (src: string) => new Promise<string[]>(async (resolve, reject) => {
+const loadThumbnails = (src: string) => new Promise<string[]>(async (resolve, reject) => {
   const canvas = document.createElement("canvas");
   const context = canvas.getContext("2d");
 
@@ -133,7 +104,7 @@ export const loadThumbnails = (src: string) => new Promise<string[]>(async (reso
   });
 });
 
-export const loadAudioBuffer = async (src: string) => new Promise<AudioBuffer>((resolve, reject) => {
+const loadAudioBuffer = async (src: string) => new Promise<AudioBuffer>((resolve, reject) => {
   const audioContext = get(studio).audioContext;
   if (!audioContext) reject("No audio context");
   else fetch(src).then(res => res.arrayBuffer())
